Increase vault init token use limit
The documentation for vault suggests creating a token from the
root token specifically for the vault authorisation action.
To lock the token down the number of times the token can be used is
set. The charm action uses the token multiple times so the action
fails. Since the token is created with a relatively short ttl it
seems reasonable to remove the usage limit entirely.

Feedback from users has shown that when following the documentation
the need to unseal each vault unit individually was missed so this
instruction is now next to the unseal explanation as well as being
reiterated in the HA section.

When deploying an HA configuration etcd must support etcd3, this
requirement was lacking in the instructions so it has been added.

Change-Id: I07c233ccadd52ec7fe2fff6276e56cd88c48acf1
Closes-Bug: #1779875