Improve upgrade guide according to bugs and tips
- When upgrading HA-clustered deployments with pause/resume,
hacluster units need to be paused first so VIP are migrated and
the unit being upgraded does not receive API requests.

- Currently nova-related charms do not support RPC version
pinning or auto-negotiation, so it is not recommended that
nova-compute is upgraded separately than nova-cloud-controller
(should be upgraded immediately after), as RPC version mismatch
will cause internal errors and API unavailability.

Change-Id: Ib63999848bae754e8fe61cd5f55345a000bc847c