clarify data-port for neutron
Update metadata.yaml snippet for neutron-gateway.

Add 'NICs' column to a table on the install-maas page.

Clarify what 'data-port' is and how to configure all MAAS nodes so that
an unused network interface is left in the 'Unconfigured' state.

It is odd to have the user do this but the alternative was to be invasive in
terms of changes, which would result in a high risk of breakage later on. This
entire guide will be revisited and reviewed at a later date.

Closes-Bug: #1833625

Change-Id: I373eb238f10ab31ee43fdea251747393bc95b44b