Review installing OpenStack from a bundle
I don't mind moving the majority of the bundle's
README to this guide instead but I don't want to
duplicate instructions.

The configure aspect of this is also awkward as
the README provides network configuration via
(currently broken) scripts (see https://is.gd/IFCo91).
Should we forgo that and force people to use the
guide's configure page? Will that apply cleanly
to this bundle's resultant cloud? See this bug
for related brokenness: https://pad.lv/1833709.

Closes-Bug: #1833533

Change-Id: Iab0807cd5079a030a4dc367366713ef04227ce42