Review deploy guide MAAS page
This is the second of an intended comprehensive series
of reviews for the deploy guide.

The idea is to pare down the amount of pure MAAS
documentation and defer mostly to upstream for details.
This will provide relief in terms of maintaining this
documentation and make this page more clear and
streamlined.

Closes-Bug: #1787192

Change-Id: I217e06f8f40f024825ee4506c29f828d45d37f73