Review deploy guide install OpenStack page
This is the fourth of an intended comprehensive series
of reviews for the deploy guide.

A correction was made to how Ceph was configured as the
backend for Cinder. The 'cinder-ceph' charm is now used.

Some drive-by improvements re formatting in recently
reviewed files.

IP addresses have been obfuscated.

The admin password is now generated.

Change-Id: I259e5349c835d67238b7b3c592246039464e4f41