Update RBD mirroring section
It was team-agreed to remove the single model
scenario. A note was left behind stating this
as an optional topology.

Move over advanced usage from the ceph-rbd-mirror
charm README. The latter content will be removed
in an imminent PR.

Removed reference to test bundles because those
bundles only cover the single-model scenario. An
overlay is also required for the two-model scenario.

Change-Id: I22e69ed67720dd98f2005b48e264ad05fa5c2573