Redirect links for config-openstack
Adds a redirect capabilities for the charm-deployment-guide for the
config-openstack.html to configure-openstack.html renaming. This also
adds whereto testing in order to validate and confirm that the redirect
links are working as intended.

Note, this will redirect the user from the config-openstack.html for any
of the release names. This may result in a 404 not found if an older
release is being accessed since the renaming/refactoring will not be
backported.

Co-Authored-By: Peter Matulis <peter.matulis@canonical.com>
Co-Authored-By: Billy Olsen <billy.olsen@canonical.com>

Change-Id: I61d2eabadd76c86fa63b49c6b7552bcb1ee20d67