Create NFV page from nova-compute README
To streamline the nova-compute charm README the
latter's NFV section is being moved to the CDG.

This PR copies the contents over wholesale.
Changes were only made to make the new page work
in its new location.

When merged, the charm's README contents will be
replaced by a summary of NFV support and linked to
the new page.

Change-Id: I220ad1a178a253ca04e1667e8c5727f8a54aa5cd