Remove migrated operations
The content of the Operations section was recently
migrated to the Admin Guide (Charm Guide) so it can
now be removed from this guide.

Add a temporary (one cycle) breadcrumb in the now-empty
Operations section so regular readers are not left
stranded. I tried to put the visible link in italics
but it is does not seem possible.

Add HTTP redirects as needed.

Change-Id: Iaf14fd71e460e607311a0a3ff2b2b6d3bc781a36