Clarify and streamline upgrade methods
Clarify - take a less prescribed approach as to
which of the three payload upgrade methods to use.
Allow the operator to choose their poison according
to their needs and desires.

Streamline - reduce the chance of user-headscratching
by removing information of a redundant or dubious
nature.

This PR is a retake of an abandoned PR:

Ie033471930e8cd2f7cb4ba110a92ff6897ae15c9

Closes-Bug: #1901592
Change-Id: Ia62695cdef916a8f78de0ccae04074fc66ae7dbd