Fix up order of charm and payload upgrades
Improve the information on order lists for both charm and
payload upgrades.

Add the following charms to the charms order:

* manila-ganesha
* octavia

Add the following charm to the payload order:

* manila-ganesha

The manila-generic alpha charm was erroneously included
in both lists. Remove it.

Add a note to both pages stating that only stable charms
are currently considered.

On the charms page it is mentioned that the given order is
based on internal testing because it is also stated that
order is irrelevant. The sentence was dropped from the
payload page because the order is not irrelevant.

Closes-Bug: #1959632
Change-Id: I3a71c3ed1a1d5c0463db8ca3ec598cd53aedf3c7