Fix up upgrade order
Correct a faulty comparison between the neutron-gateway
and ovn-chassis charms. The latter is a subordinate
charm. Bring up ovn-dedicated-chassis from the bottom
of the list.

Place nova-compute directly after ovn-central to
reduce OVN control plane downtime. Add an explanatory
admonishment.

Reflect all the above in both the upgrade-charms and
upgrade-openstack pages.

Due to the connection between this documentation and
internal testing, a comment will be added to an existing
zaza-openstack-tests ordering bug (GH #709) explaining
what was done here.

Drive-by: Implement a more efficient build invocation
          by taking advantage of multi-core systems in
          addition to only building changed files.

Related-Bug: #1960275
Change-Id: I18c65bfc333912a480e7942a06641fce3971834e