Fix series upgrade guiding principle
The series upgrade guiding principle needs to change
the order of leader vs non-leader unit machines. This
corresponds to what is in the upgrade-series-openstack
page.

Edit the associated examples accordingly.

Remove a warning about a trusty:xenial upgrade scenario
as this is now considered obsolete.

Minor word refactor and improvements.

Change-Id: Ic3cb50e2afdcd80b2d7c9ab03cf63096e03ca2d0