Remove security pages
Remove the Security pages as they are being migrated
to the charm-guide.

Some links need to be changed as a consequence.

Add HTML redirects (and tests).

The security overview page was completely removed
(not migrated) as its contents did not warrant a page
of its own. The little it had will be incorporated at
a later date.

Change-Id: If1bcbb39b4186fac7073f77bc0dacdac4294d3ff