Refactor issues and procedures
Centralise all noteworthy issues and procedures into a
single page. This will make it much simpler for users to
find important information and will reduce the necessity
to include multiple links in many pages.

Move the placement charm procedure to its own special charm
procedure page and link to it from the stein:train upgrade
path page.

The new page is currently in the Upgrades section but this
is temporary. It will soon be migrated to the charm-guide
where it will be placed under the Project section.

Include an HTML redirect (and test) for the former
upgrade-special page.

Change-Id: I3a8abe0d6b82447ca6d2883065cb1a2ce6d906b9