Remove bug and update versions
A Juju bug citation (and its workaround) was removed.
LP #1966664 is gnarly and is not completely resolved, but it no
longer affects this guide.

During the testing of the above updated software was used so
update the versions stated in the guide. In particular, a Jammy
Juju controller was used.

Related-Bug: #1966664
Related-Bug: #1971616
Change-Id: I0276861e672b24e0c886a64b5a04bb8e30caf45d