Pin to Juju 2.9 for the time being
The whole guide assumes Juju 2.9 as the top page states:
> The software versions used in this guide are as follows:
> * Ubuntu 22.04 LTS (Jammy) for the Juju client, Juju controller, and
>   all cloud nodes (including containers)
> * MAAS 3.3.3
> * Juju 2.9.43
> * OpenStack 2023.1 (Antelope)

Now that the default channel of Juju snap is 3.1, we should pin to 2.9
in the short term. Otherwise the first user experience is not going to
be smooth due to backward-incompatible commands such as `juju
add-relation` or `juju run`.

Closes-Bug: #2029323
Change-Id: Ia5eb711de1ef60443311413477460620de17f77d
(cherry picked from commit d78e6e79bf2cd0f46bbe5a12faa2febbe8998709)